requirements.txt
seaborn==0.10.1
requests==2.23.0
streamlit==0.61.0
numpy==1.18.3
matplotlib==3.2.1
pandas==1.0.3
Pillow==7.2.0
